[not verified] Slideshow: avoid loading editor on page view
Fixed by moving the pagination code to its own module, instead of having
it live in `slideshow.js`,  which depends on the editor.
