+++ projects/plugins/jetpack/extensions/blocks/slideshow/slideshow.js
@@ -5,7 +5,7 @@ import { isBlobURL } from '@wordpress/blob';
 import { RichText } from '@wordpress/block-editor';
 import { Spinner } from '@wordpress/components';
 import { Component, createRef } from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import classnames from 'classnames';
 import { isEqual } from 'lodash';
 import ResizeObserver from 'resize-observer-polyfill';
@@ -13,6 +13,7 @@ import ResizeObserver from 'resize-observer-polyfill';
  * Internal dependencies
  */
 import createSwiper from './create-swiper';
+import { paginationCustomRender } from './pagination';
 import {
 	swiperApplyAria,
 	swiperInit,
@@ -20,37 +21,6 @@ import {
 	swiperResize,
 } from './swiper-callbacks';
 
-export function paginationCustomRender( swiper, current, total ) {
-	let markup = '';
-
-	// Print dots pagination when total slides are less than six.
-	if ( total <= 5 ) {
-		for ( let i = 1; i <= total; i++ ) {
-			const active = i === current ? ' swiper-pagination-bullet-active' : '';
-			const cssClass = `swiper-pagination-bullet${ active }`;
-			const ariaLabel = sprintf(
-				/* translators: placeholder is the the video number to navigate to */
-				__( 'Go to slide %s', 'jetpack' ),
-				i
-			);
-
-			markup +=
-				'<button ' +
-				'class="' +
-				cssClass +
-				'" ' +
-				'tab-index="0" ' +
-				'role="button" ' +
-				'aria-label="' +
-				ariaLabel +
-				'"></button>';
-		}
-	} else {
-		markup += `<div class="swiper-pagination-simple">${ current } / ${ total }</div>`;
-	}
-
-	return markup;
-}
 class Slideshow extends Component {
 	pendingRequestAnimationFrame = null;
 	resizeObserver = null;
