+++ projects/plugins/jetpack/extensions/blocks/slideshow/view.js
@@ -2,7 +2,7 @@ import domReady from '@wordpress/dom-ready';
 import { forEach } from 'lodash';
 import ResizeObserver from 'resize-observer-polyfill';
 import createSwiper from './create-swiper';
-import { paginationCustomRender } from './slideshow';
+import { paginationCustomRender } from './pagination';
 import {
 	swiperApplyAria,
 	swiperInit,
